Move event handler protocols and related code to new file
