+++ tests/plugin_manager/test_command_sender_binding.py
@@ -11,8 +11,8 @@ from streamdeck.actions import Action
 if TYPE_CHECKING:
     from functools import partial
 
-    from streamdeck.actions import EventHandlerFunc
     from streamdeck.command_sender import StreamDeckCommandSender
+    from streamdeck.event_handlers.protocol import EventHandlerFunc
     from streamdeck.manager import PluginManager
     from streamdeck.models import events
